:truck: Rename(boardList): boardList 다시 활동 컴포넌트로 변경
설계오류로 롤백
