+++ src/components/activity/board-list-item.tsx
