+++ src/components/activity/board-list.tsx
@@ -1,4 +1,4 @@
-import { BoardListItem } from '@/components/board-list-item'
+import { BoardListItem } from '@/components/activity/board-list-item'
 
 export const BoardList = ({ boardListData }: { boardListData: Board[] }) => {
   return (
